% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_populations.R
\name{area_populations}
\alias{area_populations}
\title{A function to scrape the historic population data from wikipedia.
The function returns a dataframe.}
\usage{
area_populations(area_name, state, type = "county")
}
\description{
A function to scrape the historic population data from wikipedia.
The function returns a dataframe.
}
