% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fxr.R
\name{state_fxr}
\alias{state_fxr}
\title{Prepare state's name for wikipedia link}
\usage{
state_fxr(x)
}
\arguments{
\item{x}{A character string of a state's name}
}
\description{
Prepare state's name for wikipedia link
}
\examples{
state_fxr("New York")
}
