% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_link_chkr.R
\name{wiki_link_chkr}
\alias{wiki_link_chkr}
\title{A function to assemble and check the wikipedia link}
\usage{
wiki_link_chkr(x, area_name)
}
\description{
A function to assemble and check the wikipedia link
}
