% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_link_comp.R
\name{wiki_link_comp}
\alias{wiki_link_comp}
\title{A function to to make the components of a wikipedia link}
\usage{
wiki_link_comp(type, state)
}
\description{
A function to to make the components of a wikipedia link
}
